Fix empty desktop navigation pads
Angular does not fully remove previous scope elements before creating
new ones. This was making the <div id="pad"> to be repeated in the DOM
when navigating between projects

The first time a project is loaded, the bug doesn't show, as the time
required for SwellRT to get the project data is enough for the <div
id="pad"> to be removed.

After that, project model is cached, so there are two <div id="pad"> in
the DOM, angular-swellrt assigns the editor to the first one, but then
that element is removed from the DOM. So no editor is attached to the
right <div id="pad"> element

We are relying in angular-swellrt editor counter for pad ids, which is
enough so the ids are not repeated, even when the elements coexist in
the DOM

Fixes #289